chore: Move split panel side and bottom positions into separate files
